requirements.txt
httpx
beautifulsoup4
selenium
python-dateutil
tzdata
